Reintegrated bugfix branch (r4995 - r5498): Bugfixes to e.g. leaching, eet, BVOC; Support for new graphical shell; Improved LU/LCC, Forestry.
Summary of changes:
 * Addressing various bugs, e.g.:
   - Change to eet affects output in polar regions.
   - Change to leafT() affects isoprene and monoterpene emissions.
   - Further small bugfixes and technical improvements. See individual commit messages in the bugfix branch r4995 - r5495.
 * Corrected leaching function: gives significant effect on npp and other output. 
 * A few structural changes to the implementation of Landcover change and Forestry. Results are largely unaffected. 
   See the documentation and individual commit messages in the bugfix branch r4995 - r5495.
 * Support for the new version of the LPJ-GUESS Windows Shell, including reinstated 3D visualisation of vegetation structure removed in version 3.0. 64-bit (x64) compilation of the  model as a Windows dynamic-link library (DLL) is required. The shell binary is no longer stored in the svn repository: see the windows_version\readme_guesswin.txt file for its place of storage.
 * Fixed some of the new compiler warnings which appear only with VSx2012 x64.


git-svn-id: svn://stormbringer.nateko.lu.se/svn/LPJ-GUESS/trunk@5507 70805675-57b6-444b-bb00-18ecffcb4be8

+++ framework/outputchannel.cpp
@@ -90,7 +90,7 @@ bool Table::invalid() const {
 }
 
 Table OutputChannel::create_table(const TableDescriptor& descriptor) {
-	 Table table(table_descriptors.size());
+	 Table table((int) table_descriptors.size());
 
 	 table_descriptors.push_back(descriptor);
 	 values.resize(table_descriptors.size());
@@ -234,7 +234,7 @@ void FileOutputChannel::finish_row(const Table& table,
 		  }
 
 		  // print each column title
-		  int nbr_cols = get_table_descriptor(table).columns().size();
+		  int nbr_cols = (int) get_table_descriptor(table).columns().size();
 		  for (int i = 0; i < nbr_cols; i++) {
 				fputs(format_header(table, i), file);
 		  }
@@ -253,7 +253,7 @@ void FileOutputChannel::finish_row(const Table& table,
 
 	 // print out the values
 	 for (size_t i = 0; i < row.size(); i++) {
-		  fprintf(file, format(table, i), row[i]);
+		  fprintf(file, format(table, (int)i), row[i]);
 	 }
 	 fprintf(file, "\n");
 	 fflush(file);
